# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {"title": "Study Companion Project", "version": "0.1.0"},
  "paths": {
...
    "/api/v1/courses/": {
      "get": {
        "tags": ["courses"],
        "summary": "Read Courses",
        "description": "Retrieve courses.",
        "operationId": "courses-read_courses",
        "security": [{"OAuth2PasswordBearer": []}],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {"type": "integer", "default": 0, "title": "Skip"}
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {"type": "integer", "default": 100, "title": "Limit"}
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/CoursesPublic"}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/HTTPValidationError"}
              }
            }
          }
        }
      },
      "post": {
        "tags": ["courses"],
        "summary": "Create Course",
        "description": "Create new course.",
        "operationId": "courses-create_course",
        "security": [{"OAuth2PasswordBearer": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/CourseCreate"}
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/CoursePublic"}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/HTTPValidationError"}
              }
            }
          }
        }
      }
    },
    "/api/v1/courses/{id}": {
      "get": {
        "tags": ["courses"],
        "summary": "Read Course",
        "description": "Get course by ID, including its documents.",
        "operationId": "courses-read_course",
        "security": [{"OAuth2PasswordBearer": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {"type": "string", "format": "uuid", "title": "Id"}
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/CourseWithDocuments"}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/HTTPValidationError"}
              }
            }
          }
        }
      },
      "put": {
        "tags": ["courses"],
        "summary": "Update Course",
        "description": "Update an course.",
        "operationId": "courses-update_course",
        "security": [{"OAuth2PasswordBearer": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {"type": "string", "format": "uuid", "title": "Id"}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/CourseUpdate"}
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/CoursePublic"}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/HTTPValidationError"}
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["courses"],
        "summary": "Delete Course",
        "description": "Delete an course.",
        "operationId": "courses-delete_course",
        "security": [{"OAuth2PasswordBearer": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {"type": "string", "format": "uuid", "title": "Id"}
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Message"}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/HTTPValidationError"}
              }
            }
          }
        }
      }
    },
    "/api/v1/courses/{course_id}/documents": {
      "get": {
        "tags": ["courses"],
        "summary": "List Documents",
        "description": "List documents for a specific course.",
        "operationId": "courses-list_documents",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {"type": "string", "title": "Course Id"}
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {"type": "integer", "default": 0, "title": "Skip"}
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {"type": "integer", "default": 100, "title": "Limit"}
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {"type": "object", "additionalProperties": true},
                  "title": "Response Courses-List Documents"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/HTTPValidationError"}
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/process": {
      "post": {
        "tags": ["documents"],
        "summary": "Process Multiple Documents",
        "description": "Accept multiple PDF uploads, create a Document record for each,\nsave to temp files, and queue a background task for each.",
        "operationId": "documents-process_multiple_documents",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_documents-process_multiple_documents"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {"application/json": {"schema": {}}}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/HTTPValidationError"}
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{id}": {
      "get": {
        "tags": ["documents"],
        "summary": "Read Document",
        "description": "Get a document by its ID, ensuring the user has permissions.",
        "operationId": "documents-read_document",
        "security": [{"OAuth2PasswordBearer": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {"type": "string", "format": "uuid", "title": "Id"}
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Document"}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/HTTPValidationError"}
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_documents-process_multiple_documents": {
        "properties": {
          "files": {
            "items": {"type": "string", "format": "binary"},
            "type": "array",
            "title": "Files"
          },
          "course_id": {
            "type": "string",
            "format": "uuid",
            "title": "Course Id"
          }
        },
        "type": "object",
        "required": ["files", "course_id"],
        "title": "Body_documents-process_multiple_documents"
      },
...
      "CourseCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 3,
            "title": "Name"
          },
          "description": {
            "anyOf": [{"type": "string", "maxLength": 1020}, {"type": "null"}],
            "title": "Description"
          }
        },
        "type": "object",
        "required": ["name"],
        "title": "CourseCreate"
      },
      "CoursePublic": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 3,
            "title": "Name"
          },
          "description": {
            "anyOf": [{"type": "string", "maxLength": 1020}, {"type": "null"}],
            "title": "Description"
          },
          "id": {"type": "string", "format": "uuid", "title": "Id"},
          "owner_id": {"type": "string", "format": "uuid", "title": "Owner Id"}
        },
        "type": "object",
        "required": ["name", "id", "owner_id"],
        "title": "CoursePublic"
      },
      "CourseUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {"type": "string", "maxLength": 255, "minLength": 3},
              {"type": "null"}
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [{"type": "string", "maxLength": 1020}, {"type": "null"}],
            "title": "Description"
          }
        },
        "type": "object",
        "title": "CourseUpdate"
      },
      "CourseWithDocuments": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 3,
            "title": "Name"
          },
          "description": {
            "anyOf": [{"type": "string", "maxLength": 1020}, {"type": "null"}],
            "title": "Description"
          },
          "id": {"type": "string", "format": "uuid", "title": "Id"},
          "owner_id": {"type": "string", "format": "uuid", "title": "Owner Id"},
          "documents": {
            "items": {"$ref": "#/components/schemas/DocumentPublic"},
            "type": "array",
            "title": "Documents",
            "default": []
          }
        },
        "type": "object",
        "required": ["name", "id", "owner_id"],
        "title": "CourseWithDocuments"
      },
      "CoursesPublic": {
        "properties": {
          "data": {
            "items": {"$ref": "#/components/schemas/CoursePublic"},
            "type": "array",
            "title": "Data"
          },
          "count": {"type": "integer", "title": "Count"}
        },
        "type": "object",
        "required": ["data", "count"],
        "title": "CoursesPublic"
      },
      "Document": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "id": {"type": "string", "format": "uuid", "title": "Id"},
          "chunk_count": {
            "anyOf": [{"type": "integer"}, {"type": "null"}],
            "title": "Chunk Count"
          },
          "course_id": {
            "type": "string",
            "format": "uuid",
            "title": "Course Id"
          },
          "embedding_namespace": {
            "anyOf": [{"type": "string"}, {"type": "null"}],
            "title": "Embedding Namespace"
          },
          "filename": {"type": "string", "title": "Filename"},
          "status": {
            "$ref": "#/components/schemas/DocumentStatus",
            "default": "pending"
          },
          "uploaded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Uploaded At"
          }
        },
        "type": "object",
        "required": ["title", "course_id", "filename"],
        "title": "Document"
      },
      "DocumentPublic": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "id": {"type": "string", "format": "uuid", "title": "Id"},
          "course_id": {
            "type": "string",
            "format": "uuid",
            "title": "Course Id"
          },
          "uploaded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Uploaded At"
          },
          "status": {"$ref": "#/components/schemas/DocumentStatus"}
        },
        "type": "object",
        "required": ["title", "id", "course_id", "uploaded_at", "status"],
        "title": "DocumentPublic"
      },
      "DocumentStatus": {
        "type": "string",
        "enum": ["pending", "processing", "completed", "failed"],
        "title": "DocumentStatus"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {"$ref": "#/components/schemas/ValidationError"},
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Message": {
        "properties": {"message": {"type": "string", "title": "Message"}},
        "type": "object",
        "required": ["message"],
        "title": "Message"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {"anyOf": [{"type": "string"}, {"type": "integer"}]},
            "type": "array",
            "title": "Location"
          },
          "msg": {"type": "string", "title": "Message"},
          "type": {"type": "string", "title": "Error Type"}
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {"scopes": {}, "tokenUrl": "/api/v1/login/access-token"}
        }
      }
    }
  }
}